Partial refactor to move some stuff down into internals
