+++ project/plugins.sbt
@@ -1,2 +1,3 @@
 addSbtPlugin("ch.epfl.lamp" % "sbt-dotty" % "0.4.6")
 addSbtPlugin("com.geirsson" % "sbt-ci-release" % "1.5.4")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.6.1")
